Always fetch article
