+++ packages/help-center/src/components/help-center-article-fetching-content.tsx
@@ -1,31 +1,10 @@
 /* eslint-disable no-restricted-imports */
-import { useLocale } from '@automattic/i18n-utils';
 import { useEffect } from 'react';
-import { canAccessWpcomApis } from 'wpcom-proxy-request';
 import { SUPPORT_BLOG_ID } from '../constants';
 import { usePostByKey } from '../hooks/use-post-by-key';
-import { useSupportArticleAlternatesQuery } from '../hooks/use-support-article-alternates-query';
+import { useSupportArticleAlternatePostKey } from '../hooks/use-support-article-alternates-query';
 import ArticleContent from './help-center-article-content';
 import './help-center-article-content.scss';
-
-const getPostKey = ( blogId: number, postId: number ) => ( { blogId, postId } );
-
-const useSupportArticleAlternatePostKey = ( blogId: number, postId: number ) => {
-	const locale = useLocale();
-	const supportArticleAlternates = useSupportArticleAlternatesQuery( blogId, postId, locale, {
-		enabled: canAccessWpcomApis(),
-	} );
-	// Alternates don't work on Atomic.
-	if ( supportArticleAlternates.isInitialLoading && canAccessWpcomApis() ) {
-		return null;
-	}
-
-	if ( ! supportArticleAlternates.data ) {
-		return getPostKey( blogId, postId );
-	}
-
-	return getPostKey( supportArticleAlternates.data.blog_id, supportArticleAlternates.data.page_id );
-};
 
 interface ArticleFetchingContentProps {
 	postId: number;
