+++ packages/help-center/src/hooks/use-support-article-alternates-query.ts
@@ -1,5 +1,6 @@
+import { useLocale } from '@automattic/i18n-utils';
 import { useQuery } from '@tanstack/react-query';
-import wpcomRequest from 'wpcom-proxy-request';
+import wpcomRequest, { canAccessWpcomApis } from 'wpcom-proxy-request';
 
 export function useSupportArticleAlternatesQuery(
 	blogId: number,
@@ -23,3 +24,22 @@ export function useSupportArticleAlternatesQuery(
 		},
 	} );
 }
+
+const getPostKey = ( blogId: number, postId: number ) => ( { blogId, postId } );
+
+export const useSupportArticleAlternatePostKey = ( blogId: number, postId: number ) => {
+	const locale = useLocale();
+	const supportArticleAlternates = useSupportArticleAlternatesQuery( blogId, postId, locale, {
+		enabled: canAccessWpcomApis(),
+	} );
+	// Alternates don't work on Atomic.
+	if ( supportArticleAlternates.isInitialLoading && canAccessWpcomApis() ) {
+		return null;
+	}
+
+	if ( ! supportArticleAlternates.data ) {
+		return getPostKey( blogId, postId );
+	}
+
+	return getPostKey( supportArticleAlternates.data.blog_id, supportArticleAlternates.data.page_id );
+};
